+++ CHANGES.rst
@@ -29,6 +29,7 @@ in progress
 - [io] Fix MQTT error signalling
 - [grafana] Use specific Grafana folder for stuffing instant dashboards into
 - [grafana] Use topology information for deriving unique dashboard id from
+- [grafana] Add mechanisms for taming the dashboard refresh interval
 
 
 .. _kotori-0.20.1: